src/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.0.14
django-autoslug==1.9.8
django-filter==2.4.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
munch==2.5.0
Pillow==8.2.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.2.0
